common: Remove split between model and server types
