+++ credentialsd-common/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod client;
 pub mod memfd;
 pub mod model;
-pub mod server;
