+++ credentialsd/src/credential_service/usb.rs
@@ -13,14 +13,13 @@ use libwebauthn::{
     webauthn::{Error as WebAuthnError, WebAuthn},
     UvUpdate,
 };
-use tokio::sync::broadcast;
-use tokio::sync::mpsc::{self, Receiver, Sender, WeakSender};
+use tokio::sync::{
+    broadcast,
+    mpsc::{self, Receiver, Sender, WeakSender},
+};
 use tracing::{debug, warn};
 
-use credentialsd_common::{
-    model::{Credential, Error},
-    server::BackgroundEvent,
-};
+use credentialsd_common::model::{BackgroundEvent, Credential, Error};
 
 use crate::model::{CredentialRequest, GetAssertionResponseInternal};
